test: split codex role boundary coverage
